requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pygame-ce==2.5.1
httpx==0.27.2
pytest==8.3.3
pyinstaller==6.10.0
